+++ apps/builder/app/canvas/canvas.tsx
@@ -13,8 +13,6 @@ import * as baseComponents from "@webstudio-is/sdk-components-react";
 import * as baseComponentMetas from "@webstudio-is/sdk-components-react/metas";
 import * as baseComponentPropsMetas from "@webstudio-is/sdk-components-react/props";
 import { hooks as baseComponentHooks } from "@webstudio-is/sdk-components-react/hooks";
-import * as remixComponentMetas from "@webstudio-is/sdk-components-react-remix/metas";
-import * as remixComponentPropsMetas from "@webstudio-is/sdk-components-react-remix/props";
 import * as radixComponents from "@webstudio-is/sdk-components-react-radix";
 import * as radixComponentMetas from "@webstudio-is/sdk-components-react-radix/metas";
 import * as radixComponentPropsMetas from "@webstudio-is/sdk-components-react-radix/props";
@@ -233,8 +231,8 @@ export const Canvas = () => {
         // not related to sdk-components-react-remix anymore
         Body,
       },
-      metas: remixComponentMetas,
-      propsMetas: remixComponentPropsMetas,
+      metas: {},
+      propsMetas: {},
     });
     registerComponentLibrary({
       namespace: "@webstudio-is/sdk-components-react-radix",
