--- packages/sdk-components-react-remix/src/body.ws.ts
@@ -1,2 +1,0 @@
-export { Body as meta } from "@webstudio-is/sdk-components-react/metas";
-export { Body as propsMeta } from "@webstudio-is/sdk-components-react/props";
