--- packages/sdk-components-react-remix/src/metas.ts
@@ -1,5 +1,0 @@
-export { meta as Body } from "./body.ws";
-export { meta as Link } from "./link.ws";
-export { meta as RichTextLink } from "./rich-text-link.ws";
-export { meta as Form } from "./webhook-form.ws";
-export { meta as RemixForm } from "./remix-form.ws";
