--- packages/sdk-components-react-remix/src/props.ts
@@ -1,5 +1,0 @@
-export { propsMeta as Body } from "./body.ws";
-export { propsMeta as Link } from "./link.ws";
-export { propsMeta as RichTextLink } from "./rich-text-link.ws";
-export { propsMeta as Form } from "./webhook-form.ws";
-export { propsMeta as RemixForm } from "./remix-form.ws";
