--- packages/sdk-components-react-remix/src/remix-form.ws.ts
@@ -1,2 +1,0 @@
-export { Form as meta } from "@webstudio-is/sdk-components-react/metas";
-export { Form as propsMeta } from "@webstudio-is/sdk-components-react/props";
