--- packages/sdk-components-react-remix/src/rich-text-link.ws.ts
@@ -1,6 +1,0 @@
-import type { WsComponentPropsMeta } from "@webstudio-is/react-sdk";
-import { propsMeta as linkPropsMeta } from "./link.ws";
-
-export { RichTextLink as meta } from "@webstudio-is/sdk-components-react/metas";
-
-export const propsMeta: WsComponentPropsMeta = linkPropsMeta;
