+++ packages/sdk-components-react/src/rich-text-link.tsx
@@ -1,8 +1,1 @@
-import { forwardRef } from "react";
-import { Link } from "./link";
-
-export const RichTextLink: typeof Link = forwardRef((props, ref) => (
-  <Link {...props} ref={ref} />
-));
-
-RichTextLink.displayName = "RichTextLink";
+export { Link as RichTextLink } from "./link";
